[hotfix][FLINK-16047] Fix compile problem in GroupAggregateHarnessTest
